+++ app/src/main/java/holamundo/com/tenderos/ListaProductos.java
@@ -2,8 +2,6 @@ package holamundo.com.tenderos;
 
 import android.content.Intent;
 import android.os.Bundle;
-import android.support.design.widget.FloatingActionButton;
-import android.support.design.widget.Snackbar;
 import android.support.v7.app.AppCompatActivity;
 import android.support.v7.widget.LinearLayoutManager;
 import android.support.v7.widget.RecyclerView;
@@ -18,13 +16,14 @@ import com.google.firebase.database.ValueEventListener;
 
 import java.util.ArrayList;
 
-public class ListaProductos extends AppCompatActivity implements AdaptadorProducto.OnProductoClickListener{
+public class ListaProductos extends AppCompatActivity implements AdaptadorProducto.OnProductoClickListener {
     private RecyclerView lstProductos;
     private ArrayList<Producto> productos;
     private AdaptadorProducto adapter;
     private LinearLayoutManager llm;
     private String db = "productos";
     private DatabaseReference databaseReference;
+
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
@@ -37,7 +36,7 @@ public class ListaProductos extends AppCompatActivity implements AdaptadorProduc
 
         llm = new LinearLayoutManager(this);
         llm.setOrientation(LinearLayoutManager.VERTICAL);
-        adapter = new AdaptadorProducto(productos,this);
+        adapter = new AdaptadorProducto(productos, this);
 
         lstProductos.setLayoutManager(llm);
         lstProductos.setAdapter(adapter);
@@ -47,8 +46,8 @@ public class ListaProductos extends AppCompatActivity implements AdaptadorProduc
             @Override
             public void onDataChange(DataSnapshot dataSnapshot) {
                 productos.clear();
-                if (dataSnapshot.exists()){
-                    for (DataSnapshot snapshot:dataSnapshot.getChildren()){
+                if (dataSnapshot.exists()) {
+                    for (DataSnapshot snapshot : dataSnapshot.getChildren()) {
                         Producto p = snapshot.getValue(Producto.class);
                         productos.add(p);
                     }
@@ -63,29 +62,31 @@ public class ListaProductos extends AppCompatActivity implements AdaptadorProduc
             }
         });
     }
-    public void launchRegistrar(View v){
-        Intent i = new Intent(ListaProductos.this,AgregarProducto.class);
+
+    public void launchRegistrar(View v) {
+        Intent i = new Intent(ListaProductos.this, AgregarProducto.class);
         startActivity(i);
     }
+
     public void onBackPressed() {
         finish();
-        Intent i = new Intent(ListaProductos.this,Tendero.class);
+        Intent i = new Intent(ListaProductos.this, Tendero.class);
         startActivity(i);
     }
 
     @Override
     public void onProductoClickListener(Producto p) {
-        Intent i = new Intent(ListaProductos.this,DetalleProducto.class);
+        Intent i = new Intent(ListaProductos.this, DetalleProducto.class);
         Bundle b = new Bundle();
-        b.putString("id",p.getId());
-        b.putString("nombre",p.getNombre());
-        b.putString("tipo",p.getTipo());
-        b.putString("medida",p.getUnidadDeMedida());
-        b.putDouble("cantidad",p.getCantidadDisponible());
-        b.putLong("precio",p.getPrecio());
-        b.putInt("foto",p.getFoto());
+        b.putString("id", p.getId());
+        b.putString("nombre", p.getNombre());
+        b.putString("tipo", p.getTipo());
+        b.putString("medida", p.getUnidadDeMedida());
+        b.putDouble("cantidad", p.getCantidadDisponible());
+        b.putDouble("precio", p.getPrecio());
+        b.putInt("foto", p.getFoto());
 
-        i.putExtra("datos",b);
+        i.putExtra("datos", b);
         startActivity(i);
 
     }
